+++ routes/web.php
@@ -388,6 +388,7 @@ Route::middleware(['auth', 'throttle.sensitive'])->group(function () {
     Route::prefix('perfil')->name('profile.')->group(function () {
         Route::get('/', [App\Http\Controllers\ProfileController::class, 'show'])->name('show');
         Route::get('/operacoes', [App\Http\Controllers\ProfileController::class, 'operacoes'])->name('operacoes');
+        Route::put('/operacoes/preferida', [App\Http\Controllers\ProfileController::class, 'updateOperacaoPreferida'])->name('operacoes-preferida');
         Route::put('/', [App\Http\Controllers\ProfileController::class, 'update'])->name('update');
         Route::post('/avatar', [App\Http\Controllers\ProfileController::class, 'updateAvatar'])->name('update-avatar');
         Route::delete('/avatar', [App\Http\Controllers\ProfileController::class, 'removeAvatar'])->name('remove-avatar');